# cmd/app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Detect-first, escalate-later web scraper with LLM processing.",
        "title": "Greenclaw API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/extract": {
            "post": {
                "description": "Scrapes the given URL and returns structured content. For YouTube URLs, also returns transcript and LLM processing results.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "extract"
                ],
                "summary": "Extract content from a URL",
                "parameters": [
                    {
                        "description": "URL to extract",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.extractRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.errorResponse"
                        }
                    }
                }
            }
        },
        "/extract/stream": {
            "post": {
                "description": "Same as /extract but streams chunk-level progress events via Server-Sent Events while the LLM processes the transcript. Final result is sent as an \"result\" event.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "extract"
                ],
                "summary": "Extract content from a URL with SSE progress streaming",
                "parameters": [
                    {
                        "description": "URL to extract",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.extractRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.errorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "server.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "server.extractRequest": {
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "url": {
                    "type": "string",
                    "example": "https://www.youtube.com/watch?v=dQw4w9WgXcQ"
                }
            }
        },
        "store.CaptionTrack": {
            "type": "object",
            "properties": {
                "auto_generated": {
                    "type": "boolean"
                },
                "language": {
                    "type": "string"
                },
                "language_code": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "store.ContentType": {
            "type": "string",
            "enum": [
                "html",
                "json",
                "xml",
                "binary",
                "youtube_video",
                "youtube_playlist",
                "youtube_channel"
            ],
            "x-enum-varnames": [
                "ContentHTML",
                "ContentJSON",
                "ContentXML",
                "ContentBinary",
                "ContentYouTubeVideo",
                "ContentYouTubePlaylist",
                "ContentYouTubeChannel"
            ]
        },
        "store.PlaylistItem": {
            "type": "object",
            "properties": {
                "index": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "video_id": {
                    "type": "string"
                }
            }
        },
        "store.ProcessingResult": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "LLM-generated JSON, shape varies by style",
                    "type": "object"
                },
                "style": {
                    "type": "string"
                }
            }
        },
        "store.Result": {
            "type": "object",
            "properties": {
                "content_type": {
                    "$ref": "#/definitions/store.ContentType"
                },
                "description": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "fetched_at": {
                    "type": "string"
                },
                "file_path": {
                    "description": "for binary downloads",
                    "type": "string"
                },
                "links": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "text": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "youtube": {
                    "$ref": "#/definitions/store.YouTubeData"
                }
            }
        },
        "store.YouTubeData": {
            "type": "object",
            "properties": {
                "audio_path": {
                    "type": "string"
                },
                "captions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/store.CaptionTrack"
                    }
                },
                "channel_id": {
                    "type": "string"
                },
                "channel_name": {
                    "type": "string"
                },
                "duration": {
                    "type": "string"
                },
                "playlist_items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/store.PlaylistItem"
                    }
                },
                "processing": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/store.ProcessingResult"
                    }
                },
                "subtitle_paths": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "upload_date": {
                    "type": "string"
                },
                "video_id": {
                    "type": "string"
                },
                "view_count": {
                    "type": "integer"
                }
            }
        }
    }
}